wip: move logs submit dialog into its own component
