+++ src/components/feedback/SendLogsDialog.tsx
@@ -1,0 +1,72 @@
+import { Dialog, DialogContent } from '@app/components/elements/dialog/Dialog.tsx';
+import { Stack } from '@app/components/elements/Stack.tsx';
+import { Typography } from '@app/components/elements/Typography.tsx';
+import { CircularProgress } from '@app/components/elements/CircularProgress.tsx';
+import { Button } from '@app/components/elements/Button.tsx';
+import { useCallback, useState } from 'react';
+import { useTranslation } from 'react-i18next';
+import { useUIStore } from '@app/store/useUIStore.ts';
+import { invoke } from '@tauri-apps/api/tauri';
+
+export function SendLogsDialog({ onSetReference }: { onSetReference?: (reference) => void }) {
+    const { t } = useTranslation(['common', 'settings'], { useSuspense: false });
+    const showLogsDialog = useUIStore((s) => s.showLogsDialog);
+    const setShowLogsDialog = useUIStore((s) => s.setShowLogsDialog);
+
+    const [error, setError] = useState('');
+    const [feedback, setFeedback] = useState('');
+    const [loading, setLoading] = useState(false);
+
+    const sendLogs = useCallback(() => {
+        setLoading(true);
+        setError('');
+        if (!feedback) {
+            setError(t('feedback-required', { ns: 'settings' }));
+            setLoading(false);
+            return;
+        }
+        invoke('send_feedback', { feedback, includeLogs: true })
+            .then((r) => {
+                setShowLogsDialog(false);
+                onSetReference?.(r);
+            })
+            .catch((error) => {
+                setError(error.toString());
+            })
+            .finally(() => {
+                setLoading(false);
+            });
+    }, [feedback, onSetReference, setShowLogsDialog, t]);
+
+    return (
+        <Dialog open={!!showLogsDialog} onOpenChange={setShowLogsDialog}>
+            <DialogContent>
+                <Stack direction="column" alignItems="center" justifyContent="space-between">
+                    <Typography variant="h3">{t('send-logs', { ns: 'settings' })}</Typography>
+                    <textarea
+                        onChange={(e) => setFeedback(e.target.value)}
+                        style={{ width: '500px', height: '500px' }}
+                        placeholder={t('your-feedback', { ns: 'settings' })}
+                    />
+                    <Typography variant={'p'} color={'red'}>
+                        {error}
+                    </Typography>
+                    <Stack direction="row">
+                        {loading ? (
+                            <CircularProgress />
+                        ) : (
+                            <>
+                                <Button disabled={loading} onClick={() => setShowLogsDialog(false)} color="warning">
+                                    {t('cancel')}
+                                </Button>
+                                <Button disabled={loading || feedback.trim() === ''} onClick={sendLogs}>
+                                    {t('submit')}
+                                </Button>
+                            </>
+                        )}
+                    </Stack>
+                </Stack>
+            </DialogContent>
+        </Dialog>
+    );
+}
