+++ src/containers/Settings/sections/general/LogsSettings.tsx
@@ -1,11 +1,7 @@
 import { Typography } from '@app/components/elements/Typography.tsx';
-import { Button } from '@app/components/elements/Button.tsx';
 import { useTranslation } from 'react-i18next';
 import { invoke } from '@tauri-apps/api/tauri';
-import { Stack } from '@app/components/elements/Stack.tsx';
-import { useCallback, useState } from 'react';
-import { Dialog, DialogContent } from '@app/components/elements/dialog/Dialog.tsx';
-import { CircularProgress } from '@app/components/elements/CircularProgress.tsx';
+import { useState } from 'react';
 import {
     SettingsGroup,
     SettingsGroupAction,
@@ -14,34 +10,15 @@ import {
     SettingsGroupWrapper,
 } from '@app/containers/Settings/components/SettingsGroup.styles.ts';
 import { ButtonBase } from '@app/components/elements/buttons/ButtonBase.tsx';
+import { SendLogsDialog } from '@app/components/feedback/SendLogsDialog.tsx';
+import { useUIStore } from '@app/store/useUIStore.ts';
 
 export default function LogsSettings() {
     const { t } = useTranslation(['common', 'settings'], { useSuspense: false });
-    const [loading, setLoading] = useState(false);
-    const [open, setOpen] = useState(false);
-    const [feedback, setFeedback] = useState('');
-    const [error, setError] = useState('');
+    const setShowLogsDialog = useUIStore((s) => s.setShowLogsDialog);
+
     const [reference, setReference] = useState('');
-    const sendLogs = useCallback(() => {
-        setLoading(true);
-        setError('');
-        if (!feedback) {
-            setError(t('feedback-required', { ns: 'settings' }));
-            setLoading(false);
-            return;
-        }
-        invoke('send_feedback', { feedback, includeLogs: true })
-            .then((r) => {
-                setOpen(false);
-                setReference(r);
-            })
-            .catch((error) => {
-                setError(error.toString());
-            })
-            .finally(() => {
-                setLoading(false);
-            });
-    }, [feedback, t]);
+
     const openLogsDirectory = () => {
         invoke('open_log_dir')
             .then(() => {
@@ -51,10 +28,6 @@ export default function LogsSettings() {
                 console.error(error);
             });
     };
-
-    const handleClose = useCallback(() => {
-        setOpen(false);
-    }, [setOpen]);
 
     return (
         <SettingsGroupWrapper>
@@ -68,38 +41,12 @@ export default function LogsSettings() {
 
                 <SettingsGroupAction>
                     <ButtonBase onClick={openLogsDirectory}>{t('open-logs-directory', { ns: 'settings' })}</ButtonBase>
-                    <ButtonBase onClick={() => setOpen(true)}>{t('send-logs', { ns: 'settings' })}</ButtonBase>
+                    <ButtonBase onClick={() => setShowLogsDialog(true)}>
+                        {t('send-logs', { ns: 'settings' })}
+                    </ButtonBase>
                 </SettingsGroupAction>
             </SettingsGroup>
-            <Dialog open={open} onOpenChange={setOpen}>
-                <DialogContent>
-                    <Stack direction="column" alignItems="center" justifyContent="space-between">
-                        <Typography variant="h3">{t('send-logs', { ns: 'settings' })}</Typography>
-                        <textarea
-                            onChange={(e) => setFeedback(e.target.value)}
-                            style={{ width: '500px', height: '500px' }}
-                            placeholder={t('your-feedback', { ns: 'settings' })}
-                        />
-                        <Typography variant={'p'} color={'red'}>
-                            {error}
-                        </Typography>
-                        <Stack direction="row">
-                            {loading ? (
-                                <CircularProgress />
-                            ) : (
-                                <>
-                                    <Button disabled={loading} onClick={handleClose} color="warning">
-                                        {t('cancel')}
-                                    </Button>
-                                    <Button disabled={loading || feedback.trim() === ''} onClick={sendLogs}>
-                                        {t('submit')}
-                                    </Button>
-                                </>
-                            )}
-                        </Stack>
-                    </Stack>
-                </DialogContent>
-            </Dialog>
+            <SendLogsDialog onSetReference={setReference} />
         </SettingsGroupWrapper>
     );
 }
