+++ src/theme/GlobalStyle.ts
@@ -17,6 +17,7 @@ export const GlobalReset = createGlobalStyle`
             outline: none;
         }
     }
+    textarea,
     input{
         all: unset;
         /* Chrome, Safari, Edge, Opera */
